Move generic functions into common.py
