+++ scripts/ls_usgs_sr_l2.py
@@ -10,15 +10,10 @@ from xml.etree import ElementTree
 
 import click
 # image boundary imports
-import rasterio
-import rasterio.features
-import rasterio.warp
-import shapely.affinity
-import shapely.geometry
-import shapely.ops
 import yaml
 from osgeo import osr
-from rasterio.errors import RasterioIOError
+
+from scripts.common import get_grid_spatial_projection
 
 ls8_images = {
     'sr_band1': 'coastal_aerosol',
@@ -50,67 +45,6 @@ ls5_7_images = {
 }
 
 
-def safe_valid_region(images, mask_value=None):
-    try:
-        return valid_region(images, mask_value)
-    except (OSError, RasterioIOError):
-        return None
-
-
-def valid_region(images, mask_value=None):
-    mask = None
-
-    for fname in images:
-        # ensure formats match
-        with rasterio.open(str(fname), 'r') as ds:
-            transform = ds.transform
-            img = ds.read(1)
-
-            if mask_value is not None:
-                new_mask = img & mask_value == mask_value
-            else:
-                new_mask = img != ds.nodata
-            if mask is None:
-                mask = new_mask
-            else:
-                mask |= new_mask
-
-    shapes = rasterio.features.shapes(mask.astype('uint8'), mask=mask)
-    shape = shapely.ops.unary_union([shapely.geometry.shape(shape) for shape, val in shapes if val == 1])
-
-    # convex hull
-    geom = shape.convex_hull
-
-    # buffer by 1 pixel
-    geom = geom.buffer(1, join_style=3, cap_style=3)
-
-    # simplify with 1 pixel radius
-    geom = geom.simplify(1)
-
-    # intersect with image bounding box
-    geom = geom.intersection(shapely.geometry.box(0, 0, mask.shape[1], mask.shape[0]))
-
-    # transform from pixel space into CRS space
-    geom = shapely.affinity.affine_transform(geom, (transform.a, transform.b, transform.d,
-                                                    transform.e, transform.xoff, transform.yoff))
-
-    output = shapely.geometry.mapping(geom)
-    output['coordinates'] = _to_lists(output['coordinates'])
-    return output
-
-
-def _to_lists(x):
-    """
-    Returns lists of lists when given tuples of tuples
-    """
-    if isinstance(x, tuple):
-        return [_to_lists(el) for el in x]
-
-    return x
-
-
-# END IMAGE BOUNDARY CODE
-
 def satellite_ref(sat):
     """
     To load the band_names for referencing either LANDSAT8 or LANDSAT7 or LANDSAT5 bands
@@ -125,19 +59,6 @@ def satellite_ref(sat):
     if r is None:
         raise ValueError("Landsat Error")
     return r
-
-
-def get_projection(fname):
-    with rasterio.open(fname, 'r') as img:
-        left, bottom, right, top = img.bounds
-        spatial_reference = str(str(getattr(img, 'crs_wkt', None) or img.crs.wkt))
-        geo_ref_points = {
-            'ul': {'x': left, 'y': top},
-            'ur': {'x': right, 'y': top},
-            'll': {'x': left, 'y': bottom},
-            'lr': {'x': right, 'y': bottom},
-        }
-        return geo_ref_points, spatial_reference
 
 
 def get_coords(geo_ref_points, spatial_ref):
@@ -181,7 +102,7 @@ def prep_dataset(path, metadata):
                    if os.path.exists(os.path.join(path, file))}
 
     sample_file = image_files['blue']
-    geo_ref_points, spatial_ref = get_projection(os.path.join(path, sample_file))
+    geo_ref_points, spatial_ref = get_grid_spatial_projection(os.path.join(path, sample_file))
     doc = {
         'id': str(uuid.uuid4()),
         'processing_level': str(level),
